backend/requirements.txt
Django==5.0.8
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
psycopg[binary]==3.2.13
redis==5.0.8
celery==5.4.0
django-cors-headers==4.4.0
python-dotenv==1.0.1
